% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_satin_app}
\alias{create_satin_app}
\title{Create a Rstudio Project bootstraping a Satin Shiny App}
\usage{
create_satin_app(path, ...)
}
\arguments{
\item{path}{path to create}

\item{...}{not used}
}
\description{
Create a Rstudio Project bootstraping a Satin Shiny App
}
