% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar.R
\name{create_sidebar}
\alias{create_sidebar}
\title{Create sidebar}
\usage{
create_sidebar(config, data_global, display_tab = function(x) {     TRUE
  })
}
\arguments{
\item{config}{Dashboard configuration read from _site.yml}

\item{data_global}{Global data passed to the sidebar}
}
\description{
Create sidebar
}
