% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{create_tab_items}
\alias{create_tab_items}
\title{Create tab items}
\usage{
create_tab_items(config, data_global, display_tab = function(x) {    
  TRUE })
}
\arguments{
\item{config}{Dashboard configuration read from _site.yml}

\item{data_global}{Global data passed to all modules}
}
\description{
Create tab items
}
