% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{do_call_2}
\alias{do_call_2}
\title{A version of do.call that throws away extraneous arguments}
\usage{
do_call_2(what, args, ...)
}
\description{
A version of do.call that throws away extraneous arguments
}
\examples{
do_call_2(rnorm, list(n = 10, mean = 0, foo = 'bar'))
}
