% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{get_module_ui}
\alias{get_module_ui}
\title{Get the ui function for a module}
\usage{
get_module_ui(., quietly = FALSE)
}
\description{
Get the ui function for a module
}
\keyword{->}
\keyword{1.}
\keyword{2.}
\keyword{3.}
\keyword{It}
\keyword{Message}
\keyword{Module}
\keyword{No}
\keyword{Placeholder}
\keyword{UI}
\keyword{be}
\keyword{been}
\keyword{cannot}
\keyword{cases:}
\keyword{found}
\keyword{function}
\keyword{handles}
\keyword{has}
\keyword{internal}
\keyword{is}
\keyword{module}
\keyword{specified}
\keyword{three}
\keyword{use}
\keyword{with}
