% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{preview_component}
\alias{preview_component}
\title{Preview a UI component in a ShinyDashboard}
\usage{
preview_component(component, title = "Preview", use_box = TRUE, ...)
}
\arguments{
\item{component}{A UI component to display in a shinydashboard.}

\item{title}{A string indicating the title to display.}

\item{use_box}{A boolean indicating if the component should be wrapped
inside a box.}

\item{...}{Additional arguments to pass on to \code{preview_module}}
}
\description{
Preview a UI component in a ShinyDashboard
}
\examples{
\dontrun{
 ui <- DT::datatable(mtcars, width = '100\%', extension = 'Responsive')
 preview_component(ui)
}
}
\seealso{
Other preview: 
\code{\link{preview_datatable}()},
\code{\link{preview_module}()}
}
\concept{preview}
