% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{preview_datatable}
\alias{preview_datatable}
\title{Preview data in a datatable}
\usage{
preview_datatable(data, fun = DT::datatable, ui = NULL, ...)
}
\arguments{
\item{data}{A data frame object.}

\item{fun}{A function that transforms the data into a datatable. It defaults
to \code{\link[DT]{datatable}}.}

\item{ui}{A \code{shiny.tag} object to include in the UI function.}

\item{...}{Additional parameters to pass to \code{fun}.}
}
\description{
Preview data in a datatable
}
\examples{
\dontrun{
 preview_datatable(mtcars,
   style = 'bootstrap',
   width = '100\%',
   extension = 'Responsive'
 )
}
}
\seealso{
Other preview: \code{\link{preview_component}},
  \code{\link{preview_module}}
}
\concept{preview}
