% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{sb_add_component}
\alias{sb_add_component}
\title{Add a component module}
\usage{
sb_add_component(module, text = module,
  dir_pages = getOption("SB_DIR_COMPONENTS", "components"), ...)
}
\description{
Add a component module
}
