% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{sb_create_project}
\alias{sb_create_project}
\title{Create an Rstudio Project bootstraping a shinybones dashboard.}
\usage{
sb_create_project(path, ...)
}
\arguments{
\item{path}{path to create}

\item{...}{not used}
}
\description{
Create an Rstudio Project bootstraping a shinybones dashboard.
}
