% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml.R
\name{sb_pin_get}
\alias{sb_pin_get}
\title{Internal function for retrieving a pin that may have a board}
\usage{
sb_pin_get(pin_name)
}
\arguments{
\item{pin_name}{Either a pin name, or a string in the form
\code{board:pin_name}.}
}
\description{
Internal function for retrieving a pin that may have a board
}
