% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml.R
\name{sb_read_yaml}
\alias{sb_read_yaml}
\title{Read a YAML configuration file for a shinybones app}
\usage{
sb_read_yaml(config_path, data_global = list())
}
\arguments{
\item{config_path}{Path to YAML file, conventionally \code{_site.yml}.}

\item{data_global}{Global data. Variables in this can be accessed
with the \code{!global} tag.}
}
\value{
The parsed configuration file.
}
\description{
Shinybones apps have custom tags that allow R objects to be
passed to shinybones modules. Use this to parse the
\code{_site.yml} file.
}
\details{
A site configuration file supports the following tags:

\describe{
  \item{\code{!expr}}{Provide arbitrary R code to be evaluated}
  \item{\code{!pin}}{Provide a pin name that can be retrieved, or
  a string of the form \code{board:pin_name}}
  \item{\code{!global}}{An accessor for an object in \code{data_global}.}
}
}
