% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runapp.R
\name{sb_ui}
\alias{sb_ui}
\title{Create a Shiny UI function based on configuration and data}
\usage{
sb_ui(config, data_global = list())
}
\arguments{
\item{config}{The structure of the shinybones website. Generally
parsed from \code{_site.yml} using \code{\link{sb_read_yaml}}.}

\item{data_global}{Global data that gets passed to everything in the
shinybones app.}
}
\value{
A function that can be passed to the \code{ui} parameter
of \code{\link[shiny]{shinyApp}}.
}
\description{
Create a Shiny UI function based on configuration and data
}
