% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{source_dirs}
\alias{source_dirs}
\title{Source R files from multiple directories recursively}
\usage{
source_dirs(..., recursive = FALSE)
}
\description{
Source R files from multiple directories recursively
}
