% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snippets.R
\name{st_add_snippets}
\alias{st_add_snippets}
\title{Add snippets to your ~/.R/snippets/r.snippets}
\usage{
st_add_snippets()
}
\description{
Add snippets to your ~/.R/snippets/r.snippets
}
\examples{
\dontrun{
  add_snippets()
}
}
