% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{st_call_all_modules}
\alias{st_call_all_modules}
\title{Call all modules}
\usage{
st_call_all_modules(config, data_global, display_tab = function(x) {    
  FALSE })
}
\arguments{
\item{config}{Dashboard configuration read from _site.yml}

\item{data_global}{Global data passed to the sidebar}
}
\description{
Call all modules
}
