% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar.R
\name{st_create_sidebar}
\alias{st_create_sidebar}
\title{Create sidebar}
\usage{
st_create_sidebar(config, data_global, display_page = function(x) {    
  TRUE })
}
\arguments{
\item{config}{Dashboard configuration read from _site.yml}

\item{data_global}{Global data passed to the sidebar.}

\item{display_page}{A function that returns a boolean indicating if a page
should be displayed in the sidebar.}
}
\description{
Create sidebar
}
