% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{st_create_tab_items}
\alias{st_create_tab_items}
\title{Create tab items}
\usage{
st_create_tab_items(config, data_global, display_tab = function(x) {    
  TRUE })
}
\arguments{
\item{config}{Dashboard configuration read from _site.yml}

\item{data_global}{Global data passed automatically from app.R}

\item{display_tab}{A function that returns a boolean indicating whether a tab is
to be displayed in the UI}
}
\description{
Create tab items
}
