% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{preview_component}
\alias{preview_component}
\title{Preview a component in a satin shinydashboard}
\usage{
preview_component(x, title = "Preview", use_box = TRUE, ...)
}
\description{
Preview a component in a satin shinydashboard
}
