% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{preview_module}
\alias{preview_module}
\alias{preview_component}
\title{Preview a shiny module or a UI component in a shinydashboard}
\usage{
preview_module(module_name, name = module_name, use_box = FALSE,
  preview = TRUE, ...)

preview_component(x, title = "Preview", use_box = TRUE, ...)
}
\arguments{
\item{module_name}{Name of the module}

\item{name}{Namespace to call the module}

\item{...}{Additional parameters to pass to the module}
}
\description{
Preview a shiny module or a UI component in a shinydashboard
}
\examples{
slider_text_ui <- function(id){
  ns <- NS(id)
  tagList(
    sliderInput(ns('num'), 'Enter Number', 0, 1, 0.5),
    textOutput(ns('num_text'))
  )
}
slider_text <- function(input, output, session){
   output$num_text <- renderText({input$num})
}
preview_module('slider_text')
}
