% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{preview_module}
\alias{preview_module}
\title{Preview a shiny module in a shinydashboard}
\usage{
preview_module(module, name = "module", use_box = FALSE,
  title = name, titleWidth = NULL, preview = TRUE, ...)
}
\arguments{
\item{module}{The server function of a module or its name.}

\item{name}{The id to pass to callModule while calling it.}

\item{use_box}{A boolean indicating if the ui should be wrapped in a
box}

\item{title}{The dashboard title to display.}

\item{titleWidth}{The width of the dashboard title.}

\item{preview}{A boolean indicating if the return value should be a shinyApp
object (default) or a named list with ui and server.}

\item{...}{Additional parameters to pass to the module}
}
\description{
Preview a shiny module in a shinydashboard
}
\examples{
\dontrun{
  library(shiny)
  slider_text_ui <- function(id){
    ns <- NS(id)
    tagList(
      sliderInput(ns('num'), 'Enter Number', 0, 1, 0.5),
      textOutput(ns('num_text'))
    )
  }
  slider_text <- function(input, output, session){
     output$num_text <- renderText({input$num})
  }
  preview_module(slider_text, title = 'Slider Text')
  preview_module(slider_text, title = 'Slider Text', use_box = TRUE)
}
}
\seealso{
Other preview: \code{\link{preview_component}},
  \code{\link{preview_datatable}}
}
\concept{preview}
