% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snippets.R
\name{sb_add_snippets}
\alias{sb_add_snippets}
\title{Add snippets to your ~/.R/snippets/r.snippets}
\usage{
sb_add_snippets(sn_file = "~/.R/snippets/r.snippets")
}
\arguments{
\item{sn_file}{Path to the file r.snippets}
}
\description{
Add snippets to your ~/.R/snippets/r.snippets
}
\examples{
\dontrun{
  sb_add_snippets()
}
}
