% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runapp.R
\name{sb_create_app}
\alias{sb_create_app}
\title{Run a ShinyDashboard App created using ShinyBones}
\usage{
sb_create_app(app_dir = ".", data_global = list())
}
\description{
Run a ShinyDashboard App created using ShinyBones
}
