% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar.R
\name{sb_create_sidebar_conditional_panels}
\alias{sb_create_sidebar_conditional_panels}
\title{Create conditional panels}
\usage{
sb_create_sidebar_conditional_panels(config, data_global = list())
}
\description{
Create conditional panels
}
