% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{source_dirs}
\alias{source_dirs}
\title{Source R files from multiple directories recursively}
\usage{
source_dirs(..., recursive = FALSE)
}
\arguments{
\item{...}{directories to source R scripts from}

\item{recursive}{logical. Should the listing recurse into directories?}
}
\description{
Source R files from multiple directories recursively
}
